test: split up api tests into respective folders
